fix precision warning

+++ SoftwareRenderer/MyBoard.cpp
@@ -23,7 +23,7 @@ void MyBoard::Render(HDC hdc, float dt)
 	static float a;
 	SRClearCanvas(_mainCanvas, RGB{ 0, 0, 0 });
 	a += 0.5f;
-	SRSetView(SRCreateViewAtLH(Vector3{ 10 * cos(deg2rad(a)), -10 + a * 0.1f, 10 * sin(deg2rad(a)) }, Vector3{ 0, 0, 0 }, Vector3{ 0, 1, 0 }));
+	SRSetView(SRCreateViewAtLH(Vector3{ 10 * (float)cos(deg2rad(a)), -10 + a * 0.1f, 10 * (float)sin(deg2rad(a)) }, Vector3{ 0, 0, 0 }, Vector3{ 0, 1, 0 }));
 	_cube.position = Vector3{ 0.5f, 0.5f, 0.5f };
 	SRDrawObject(_mainCanvas, _cube);
 	SRDrawWireFrame(_mainCanvas, _cube, RGB{ 0, 0, 255 });
